refactor blender dataset loader and simplify loader for unlabeled data
