+++ keypoint_detection/data/unlabeled_dataset.py
@@ -1,36 +1,40 @@
 from pathlib import Path
 from typing import List, Union
-
+import os
 import torch
+from torchvision.transforms import ToTensor
 
-from keypoint_detection.data.dataset import KeypointsDataset
+
+from keypoint_detection.data.utils import ImageDataset
 
 
-class UnlabeledKeypointsDataset(KeypointsDataset):
+class UnlabeledKeypointsDataset(ImageDataset):
     """
     Simple dataset to run inference on unlabeled data
     """
 
     def __init__(
         self,
-        json_dataset_path: str,
         image_dataset_path: str,
-        keypoint_channels: Union[List[str], str],
-        keypoint_channel_max_keypoints: Union[List[int], str],
         **kwargs,
     ):
+        super().__init__()
+        self.image_paths = os.listdir(image_dataset_path)
+        self.image_paths = [image_dataset_path  +f"/{path}" for path in self.image_paths]
 
-        super().__init__(
-            json_dataset_path, image_dataset_path, keypoint_channels, keypoint_channel_max_keypoints, **kwargs
-        )
+        self.transform = ToTensor()  # convert images to Torch Tensors
+
 
     def __getitem__(self, index):
         if torch.is_tensor(index):
             index = index.tolist()
         index = int(index)
-        image_path = Path(".") / self.image_dir / self.dataset[index]["image_path"]
 
+        image_path = self.image_paths[index]
         image = self.image_loader.get_image(image_path, index)
         image = self.transform(image)
 
         return image
+
+    def __len__(self):
+        return len(self.image_paths)
