+++ keypoint_detection/train/train.py
@@ -7,7 +7,7 @@ from pytorch_lightning.loggers import WandbLogger
 from pytorch_lightning.trainer.trainer import Trainer
 
 from keypoint_detection.data.datamodule import RandomSplitDataModule
-from keypoint_detection.data.dataset import KeypointsDataset
+from keypoint_detection.data.blender_dataset import BlenderKeypointsDataset
 from keypoint_detection.models.backbones.backbone_factory import BackboneFactory
 from keypoint_detection.models.detector import KeypointDetector
 from keypoint_detection.models.loss import LossFactory
@@ -57,7 +57,7 @@ def main(hparams: dict) -> Tuple[KeypointDetector, pl.Trainer]:
     loss = LossFactory.create_loss(**hparams)
     model = KeypointDetector(backbone=backbone, loss_function=loss, **hparams)
 
-    dataset = KeypointsDataset(**hparams)
+    dataset = BlenderKeypointsDataset(**hparams)
 
     module = RandomSplitDataModule(dataset, **hparams)
     wandb_logger = WandbLogger(
@@ -85,7 +85,7 @@ if __name__ == "__main__":
     parser = add_system_args(parser)
     parser = KeypointDetector.add_model_argparse_args(parser)
     parser = Trainer.add_argparse_args(parser)
-    parser = KeypointsDataset.add_argparse_args(parser)
+    parser = BlenderKeypointsDataset.add_argparse_args(parser)
     parser = RandomSplitDataModule.add_argparse_args(parser)
     parser = BackboneFactory.add_to_argparse(parser)
     parser = LossFactory.add_to_argparse(parser)
