+++ test/test_dataset.py
@@ -2,7 +2,8 @@ import unittest
 
 import torch
 
-from keypoint_detection.data.dataset import KeypointsDataset
+from keypoint_detection.data.blender_dataset import BlenderKeypointsDataset
+from keypoint_detection.data.unlabeled_dataset import UnlabeledKeypointsDataset
 
 from .configuration import DEFAULT_HPARAMS
 
@@ -12,7 +13,7 @@ class TestDataSet(unittest.TestCase):
         self.hparams = DEFAULT_HPARAMS
 
     def test_dataset(self):
-        dataset = KeypointsDataset(**self.hparams)
+        dataset = BlenderKeypointsDataset(**self.hparams)
         item = dataset.__getitem__(0)
         img, keypoints = item
 
@@ -25,3 +26,11 @@ class TestDataSet(unittest.TestCase):
 
         self.assertTrue(isinstance(img, torch.Tensor))
         self.assertTrue(isinstance(corner, torch.Tensor))
+
+class TestUnlabeledDataset(unittest.TestCase):
+    def test_dataset(self):
+        dataset = UnlabeledKeypointsDataset(DEFAULT_HPARAMS["image_dataset_path"]  +"/images")
+
+        self.assertEqual(len(dataset), 4)
+        img = dataset[0]
+        self.assertEqual(img.shape, (3,64,64))
