+++ test/test_dataset/coco_dataset.json
@@ -1,0 +1,150 @@
+{
+    "info": {
+        "description": "test dataset for keypoint detection framework",
+        "url": "todo",
+        "version": "0.0.1",
+        "year": 2022,
+        "contributor": "Thomas Lips",
+        "date_created": "2022/08/22"
+    },
+    "licenses": [
+        {
+            "url": "hhttps://creativecommons.org/licenses/by/2.0/",
+            "id": 1,
+            "name": "Attribution Generic License"
+        }
+    ],
+    "categories": [
+        {
+            "id": 1,
+            "name": "box",
+            "keypoints": [
+                "box_corner0, box_corner1,box_corner2, box_corner3",
+                "flap_corner0, flap_corner1,flap_corner2,flap_corner3,flap_corner4,flap_corner5,flap_corner6",
+                "flap_corner7"
+            ],
+            "skeleton": [],
+            "supercategory": "box"
+        }
+    ],
+    "images": [
+        {
+            "id": 1,
+            "license": 1,
+            "file_name": "images/0.png",
+            "width": 64,
+            "height": 64
+        },
+        {
+            "id": 2,
+            "license": 1,
+            "file_name": "images/1.png",
+            "width": 64,
+            "height": 64
+        },
+        {
+            "id": 3,
+            "license": 1,
+            "file_name": "images/2.png",
+            "width": 64,
+            "height": 64
+        },
+        {
+            "id": 4,
+            "license": 1,
+            "file_name": "images/3.png",
+            "width": 64,
+            "height": 64
+        }
+    ],
+    "annotations": [
+        {
+            "id": 1,
+            "image_id": 1,
+            "category_id": 1,
+            "keypoints": [
+                0.5514724254608154,
+                0.65,
+                2,
+                0.34643998742103577,
+                0.55,
+                2,
+                0.4485275447368622,
+                0.35,
+                2,
+                0.6535599827766418,
+                0.47,
+                2,
+                0.5648945569992065,
+                0.6,
+                2,
+                0.3836907148361206,
+                0.49,
+                2,
+                0.3338361084461212,
+                0.51,
+                2,
+                0.42549973726272583,
+                0.35,
+                2,
+                0.4768047034740448,
+                0.32,
+                2,
+                0.6589296460151672,
+                0.44,
+                2,
+                0.6447089910507202,
+                0.48,
+                2,
+                0.5539688467979431,
+                0.74,
+                2
+            ],
+            "num_keypoints": 12
+        },
+        {
+            "id": 2,
+            "image_id": 2,
+            "category_id": 1,
+            "keypoints": [
+                0.5514724254608154,
+                0.65,
+                2,
+                0.34643998742103577,
+                0.55,
+                2,
+                0.4485275447368622,
+                0.35,
+                2,
+                0.6535599827766418,
+                0.47,
+                2,
+                0.5648945569992065,
+                0.6,
+                2,
+                0.3836907148361206,
+                0.49,
+                2,
+                0.3338361084461212,
+                0.51,
+                2,
+                0.42549973726272583,
+                0.35,
+                2,
+                0.4768047034740448,
+                0.32,
+                2,
+                0.6589296460151672,
+                0.44,
+                2,
+                0.0,
+                0.0,
+                0,
+                0.0,
+                0.0,
+                0
+            ],
+            "num_keypoints": 10
+        }
+    ]
+}
