+++ test/test_detector.py
@@ -4,7 +4,7 @@ import torch
 from pytorch_lightning.loggers import WandbLogger
 
 from keypoint_detection.data.datamodule import RandomSplitDataModule
-from keypoint_detection.data.dataset import KeypointsDataset
+from keypoint_detection.data.blender_dataset import BlenderKeypointsDataset
 from keypoint_detection.models.backbones.unet import UnetBackbone
 from keypoint_detection.models.detector import KeypointDetector
 from keypoint_detection.models.loss import bce_loss
@@ -29,7 +29,7 @@ class TestHeatmapUtils(unittest.TestCase):
         self.backbone = UnetBackbone(**self.hparams)
         self.model = KeypointDetector(backbone=self.backbone, loss_function=self.loss_function, **self.hparams)
 
-        self.module = RandomSplitDataModule(KeypointsDataset(**self.hparams), **self.hparams)
+        self.module = RandomSplitDataModule(BlenderKeypointsDataset(**self.hparams), **self.hparams)
 
     def test_perfect_heatmap(self):
         loss = self.model.heatmap_loss(self.heatmaps, self.heatmaps)
@@ -50,7 +50,7 @@ class TestModel(unittest.TestCase):
         self.backbone = UnetBackbone(**self.hparams)
         self.model = KeypointDetector(backbone=self.backbone, loss_function=self.loss_function, **self.hparams)
 
-        self.module = RandomSplitDataModule(KeypointsDataset(**self.hparams), **self.hparams)
+        self.module = RandomSplitDataModule(BlenderKeypointsDataset(**self.hparams), **self.hparams)
 
     def test_shared_step_batch(self):
